+++ src/App/App.css
@@ -1,6 +1,7 @@
 * {
   margin: 0;
   padding: 0;
+  font-family: 'Raleway';
 }
 
 .App {
