--- src/App/App.test.js
@@ -1,34 +1,0 @@
-import React from 'react';
-import App from './App';
-import { shallow } from 'enzyme';
-
-describe('App', () => {
-  let wrapper;
-  beforeEach(() => {
-  	wrapper = shallow(<App />)
-  })
-  it('should match a snapshot', () => {
-  	expect(wrapper).toMatchSnapshot()
-  })
-
-  describe('mapStateToProps', () => {
-    it('should return an object with the movies and ratings array and a loading status', () => {
-      const mockState = {
-        movies: [{title: 'Parasite', rating: 9}, {title: '1917', rating: 8}],
-        ratings: [{id: 25, rating: 9}, {id: 27, rating: 4}],
-        user: {id: 123, name: 'me'},
-        loadingStatus: true,
-        someOtherProp: false
-      }
-      const expected = {
-        movies: [{title: 'Parasite', rating: 9}, {title: '1917', rating: 8}],
-        ratings: [{id: 25, rating: 9}, {id: 27, rating: 4}],
-        loadingStatus: true,
-        user: {id: 123, name: 'me'},
-      }
-      const mappedProps = mapStateToProps(mockState)
-      expect(mappedProps).toEqual(expected)
-    })
-  })
-
-});
