+++ src/ReviewForm/ReviewForm.css
@@ -2,7 +2,8 @@
   border-radius: 8px;
   height: 350px;
   width: 80%;
-  background: orange;
+  background: #328cc1;
+  margin-top: 29%;
 }
 
 .review_form-container {
